fix(error-tracking): isolate issue assignee picker
